Implement style transfer algorithm.
This commit adds a `StyleTransfer` class that implements the NST
algorithm. An object of this class represents a reifeid run of
style transfer for a given content and style image pair. Calling the
`step()` function on this object results in one step of optimization
(using Adam). The current generated image can be obtained via an
`image` property.

Also, the paper in the README has been updated to the 2016 version.